start thinking about parameterising model by constraints
writing description of what the context state should look like, but this
is incomplete. thinking it might be better to base everything on
init_ctxt and treat the Booleans just as other definitions, and select
and ind just as other constraints. (the theorems proving the consistency
of the three actual axioms might need to be slightly generalised for
this usage. not sure how yet.)