some hacks for build_ConstDef
when proving term_ok it needs to check that a constant's type is an
instance of the type in the signature. the way things are currently set
up, too much information is lost to do this properly with something like
match_type, so just add a few theorems for the cases that end up not
getting proved so far.