make lcaCtxt into a lib
it is not a theory since it's main job is not to store theorems, but to
build an outer context value. unfortunately there is no nice way to
cache the result of running this lib (which takes a long time!) - one
way to do it is to use PolyML's saveState facility to build an image
that includes the resulting context value.