make termsem_cert build a valuation too
it's actually pretty simple since only type variables are relevant.

now the certificate theorems use a new constant "a_valuation" for the
valuation. nothing is proved about the term part of a_valuation, but
that shouldn't matter since the terms passed to termsem_cert are
supposed to be closed. for the type part, the valuation is (K boolset)
=++ [mappings for the relevant type variables].