Create axiomatic context and type assignment from constraints
Proved that the resulting context extends hol_ctxt, and the resulting
type assignment is a (well-formed) type assignment. Still to prove: that
it sends constrained types to the right values. Still to do: define how
to create a term assignment from similar constraints, and prove good
things about it too.