Work on other pieces of build_axiomatic_interpretation
The main missing piece now is proving models_thm, in particular, that
the ax_int actually gives the correct semantics to all the axioms. This
will require going via reflection of the truth from outer HOL.

There is also a small missing piece to do with select (int assumptions
about select will currently not be proved). This might require more
inputs to the automation (about the incoming base interpretation).